Fix daily-CI flakiness and make any tests directory run standalone
Three of the daily failures were the same disease — Panel/Holoviews global
state leaking across tests under xdist parallelism and collection order:

- Isolate the UI live Panel server into its own serial CI pass so it never
  shares pn.state with headless build_homepage() calls on the same worker
  (fixes the flaky test_consistency[artifacts_concatenate] session_context race).
- Register hv.extension("bokeh") in the session panel_extension fixture so the
  dashboard/plotter unit tests get a deterministic backend regardless of order
  (fixes pytest tests/unit erroring on TestVisualizationDashboard).
- Rewrite the progress-poller tests to synchronize on observed file output
  instead of wall-clock sleeps.

Restructure test imports so any directory runs standalone: move shared
constants/helpers out of conftest.py into guppy_test_data.py (unique name,
no conftest collision), split integration-only helpers into
integration_helpers.py, set pythonpath=["tests"], and normalize __init__.py.
