+++ .github/workflows/run-tests.yml
@@ -75,9 +75,13 @@ jobs:
           rclone-config: ${{ secrets.RCLONE_CONFIG }}
 
       - name: Run tests
-        run: pytest tests -vv -rsx -n auto --dist worksteal -m "not parallel and not dandi_live${{ inputs.skip-full-data-tests && ' and not full_data' || '' }}${{ runner.os == 'Windows' && ' and not progress_bar' || '' }}" --cov=guppy --cov-branch --cov-report=xml
+        run: pytest tests -vv -rsx -n auto --dist worksteal -m "not parallel and not ui and not dandi_live${{ inputs.skip-full-data-tests && ' and not full_data' || '' }}${{ runner.os == 'Windows' && ' and not progress_bar' || '' }}" --cov=guppy --cov-branch --cov-report=xml
       - name: Run parallel tests
         run: pytest tests -vv -rsx -m "parallel and not full_data and not dandi_live" --cov=guppy --cov-append --cov-report=xml
+      # UI tests run serially in their own process so the threaded live Panel server never shares
+      # pn.state global state with the headless build_homepage() calls in the rest of the suite.
+      - name: Run UI tests
+        run: pytest tests -vv -rsx -m "ui and not dandi_live" --cov=guppy --cov-append --cov-report=xml
       - name: Upload full coverage to Codecov
         if: ${{ matrix.python-version == '3.10' && matrix.os == 'ubuntu-latest' }}
         uses: codecov/codecov-action@v5
