+++ tests/consistency/test_consistency_artifact_removal.py
@@ -4,10 +4,10 @@ import shutil
 
 import numpy as np
 import pytest
-from conftest import TESTING_DATA, event_ts_offset_for
 
 from guppy.testing import compare_output_folders
 from guppy.testing.api import step1, step2, step3, step4
+from guppy_test_data import TESTING_DATA, event_ts_offset_for
 
 # Artifact window coordinates (x/time values only; col 1 is y and not used by the analysis).
 # Each array has shape (4, 2): four click events defining two artifact windows.
