+++ tests/integration/conftest.py
@@ -1,7 +1,6 @@
 import glob
 import os
 import shutil
-from pathlib import Path
 from unittest.mock import patch
 
 import holoviews as hv
@@ -10,80 +9,9 @@ import pytest
 from guppy.frontend.visualization_dashboard import VisualizationDashboard
 from guppy.testing.api import step1, step2, step3, step4, step5
 from guppy.utils.utils import parse_run_name
+from guppy_test_data import STUBBED_TESTING_DATA
 
-PROJECT_ROOT = Path(__file__).resolve().parents[2]
-STUBBED_TESTING_DATA = PROJECT_ROOT / "stubbed_testing_data"
-
-
-REPRESENTATIVE_SESSIONS = {
-    "csv": {
-        "session_subdir": "csv/sample_data_csv_1",
-        "store_id_to_store_label": {
-            "Sample_Control_Channel": "control_region",
-            "Sample_Signal_Channel": "signal_region",
-            "Sample_TTL": "ttl",
-        },
-        "npm_timestamp_column_names": None,
-        "npm_time_units": None,
-        "npm_split_events": [True, True],
-    },
-    "tdt": {
-        "session_subdir": "tdt/Photo_63_207-181030-103332",
-        "store_id_to_store_label": {
-            "Dv1A": "control_dms",
-            "Dv2A": "signal_dms",
-            "PrtN": "port_entries_dms",
-        },
-        "npm_timestamp_column_names": None,
-        "npm_time_units": None,
-        "npm_split_events": [True, True],
-    },
-    "npm": {
-        "session_subdir": "npm/sampleData_NPM_5",
-        "store_id_to_store_label": {
-            "file0_chev1": "control_region1",
-            "file0_chod1": "signal_region1",
-            "event0": "ttl_region1",
-        },
-        "npm_timestamp_column_names": None,
-        "npm_time_units": None,
-        "npm_split_events": None,
-    },
-    "doric": {
-        "session_subdir": "doric/sample_doric_1",
-        "store_id_to_store_label": {
-            "AIn-1 - Raw": "control_region",
-            "AIn-2 - Raw": "signal_region",
-            "DI--O-1": "ttl",
-        },
-        "npm_timestamp_column_names": None,
-        "npm_time_units": None,
-        "npm_split_events": [True, True],
-    },
-    "nwb": {
-        "session_subdir": "nwb/mock_nwbfile_ndx_fiber_photometry_v0_2_ndx_events_v0_2",
-        "store_id_to_store_label": {
-            "fiber_photometry_response_series_0": "control_region",
-            "fiber_photometry_response_series_1": "signal_region",
-            "events": "ttl",
-        },
-        "npm_timestamp_column_names": None,
-        "npm_time_units": None,
-        "npm_split_events": [True, True],
-    },
-}
-
-
-def _locate_output_directory(*, session_copy: str) -> str:
-    session_name = os.path.basename(session_copy)
-    output_directories = sorted(glob.glob(os.path.join(session_copy, f"{session_name}_output_*")))
-    assert output_directories, f"No output directories found in {session_copy}"
-
-    for output_directory in output_directories:
-        if os.path.exists(os.path.join(output_directory, "storesList.csv")):
-            return output_directory
-
-    raise AssertionError(f"No storesList.csv found in any output directory under {session_copy}")
+from .integration_helpers import REPRESENTATIVE_SESSIONS, _locate_output_directory
 
 
 def _prepare_pipeline_state(
