+++ tests/integration/test_integration_combine_data.py
@@ -6,10 +6,10 @@ from unittest.mock import patch
 import h5py
 import holoviews as hv
 import pytest
-from conftest import STUBBED_TESTING_DATA
 
 from guppy.frontend.visualization_dashboard import VisualizationDashboard
 from guppy.testing.api import step1, step2, step3, step4, step5
+from guppy_test_data import STUBBED_TESTING_DATA
 
 
 @pytest.mark.filterwarnings("ignore::UserWarning")
