+++ tests/integration/test_integration_group_analysis.py
@@ -7,10 +7,10 @@ from unittest.mock import patch
 import holoviews as hv
 import pandas as pd
 import pytest
-from conftest import STUBBED_TESTING_DATA
 
 from guppy.frontend.visualization_dashboard import VisualizationDashboard
 from guppy.testing.api import step1, step2, step3, step4, step5
+from guppy_test_data import STUBBED_TESTING_DATA
 
 SESSION_SUBDIRS = [
     "tdt/Photo_048_392-200728-121222",
