+++ tests/integration/test_integration_inter_session_mixed_modality.py
@@ -4,9 +4,9 @@ import shutil
 
 import h5py
 import pytest
-from conftest import STUBBED_TESTING_DATA
 
 from guppy.testing.api import step1, step2, step3, step4
+from guppy_test_data import STUBBED_TESTING_DATA
 
 
 @pytest.mark.filterwarnings("ignore::UserWarning")
