+++ tests/integration/test_integration_parallel.py
@@ -15,9 +15,9 @@ import shutil
 import h5py
 import pandas as pd
 import pytest
-from conftest import STUBBED_TESTING_DATA
 
 from guppy.testing.api import step1, step2, step3, step4
+from guppy_test_data import STUBBED_TESTING_DATA
 
 # Use the CSV sample as it is the simplest format — no binary TDT dependencies.
 SESSION_SUBDIR = "csv/sample_data_csv_1"
