+++ tests/unit/orchestration/test_read_raw_data.py
@@ -6,12 +6,12 @@ import threading
 import time
 
 import pytest
-from conftest import STUBBED_TESTING_DATA
 
 from guppy.orchestration.read_raw_data import (
     _progress_poller,
     orchestrate_read_raw_data,
 )
+from guppy_test_data import STUBBED_TESTING_DATA
 
 # orchestrate_read_raw_data writes the parameter snapshot (save_parameters) into each
 # selected output dir before reading, so it needs the full analysis-parameter set.
@@ -85,6 +85,24 @@ class TestOrchestrateReadRawDataErrorEnrichment:
         assert "DI--O-1" in message
 
 
+def _read_progress_values(progress_file) -> list[int]:
+    # The poller creates the file only on its first write, so callers polling for output
+    # may look before it exists.
+    if not progress_file.exists():
+        return []
+    return [int(line.strip()) for line in progress_file.read_text().splitlines() if line.strip()]
+
+
+def _wait_until(predicate, *, timeout: float = 5.0, interval: float = 0.01) -> bool:
+    """Poll ``predicate`` until it is truthy or ``timeout`` elapses. Returns whether it became true."""
+    deadline = time.monotonic() + timeout
+    while time.monotonic() < deadline:
+        if predicate():
+            return True
+        time.sleep(interval)
+    return False
+
+
 class TestProgressPoller:
     """``_progress_poller`` flushes the shared samples counter to ``PB_STEPS_FILE``."""
 
@@ -99,18 +117,21 @@ class TestProgressPoller:
         )
         thread.start()
         try:
-            # Drive three distinct values past the poller; sleep long enough between
-            # each so the 200ms poll loop sees them as separate ticks.
-            for value in (5, 12, 30):
+            # Drive three distinct values through the poller. Synchronize on the poller's
+            # observed output — wait until each value*10 has landed before driving the next —
+            # so the assertion never depends on wall-clock poll timing. The poller's first read
+            # is gated behind a 200ms wait, so value 5 is always in place before it reads.
+            for value, expected in ((5, 50), (12, 120), (30, 300)):
                 with samples_done.get_lock():
                     samples_done.value = value
-                time.sleep(0.3)
+                assert _wait_until(
+                    lambda expected=expected: _read_progress_values(progress_file)[-1:] == [expected]
+                ), f"poller never wrote {expected}; file so far: {_read_progress_values(progress_file)}"
         finally:
             stop_event.set()
             thread.join(timeout=2.0)
 
-        written_values = [int(line.strip()) for line in progress_file.read_text().splitlines() if line.strip()]
-        assert written_values == [50, 120, 300]
+        assert _read_progress_values(progress_file) == [50, 120, 300]
 
     def test_skips_redundant_writes_when_value_unchanged(self, tmp_path):
         progress_file = tmp_path / "pb.txt"
@@ -122,14 +143,20 @@ class TestProgressPoller:
             kwargs={"file_path": str(progress_file)},
         )
         thread.start()
-        time.sleep(0.7)  # ~3 poll cycles with no value change
-        stop_event.set()
-        thread.join(timeout=2.0)
+        try:
+            # Wait for the first (and only expected) write, then let several more poll
+            # cycles elapse. The value never changes, so no second write may ever appear.
+            assert _wait_until(
+                lambda: _read_progress_values(progress_file) == [70]
+            ), f"poller never wrote 70; file: {_read_progress_values(progress_file)}"
+            time.sleep(0.6)  # ~3 more poll cycles with no value change
+        finally:
+            stop_event.set()
+            thread.join(timeout=2.0)
 
-        written_values = [int(line.strip()) for line in progress_file.read_text().splitlines() if line.strip()]
         # The first poll writes the value once; subsequent polls with the same value
         # must not write again.
-        assert written_values == [70]
+        assert _read_progress_values(progress_file) == [70]
 
 
 class TestProgressFileAccountingEndToEnd:
